do async staff * Worked on subscription ( event will be fired if the subscription was modified, so a new loop will start earlier. * fixed tests * AutoResetEvent  use Async

+++ Papper/src/Papper/Extensions/Notification/Subscription.cs
@@ -11,16 +11,18 @@ namespace Papper.Extensions.Notification
     /// <summary>
     /// This representates a subscription to a plc value change detection.
     /// </summary>
-    public sealed class Subscription : IDisposable
+    public sealed class Subscription : IDisposable, IAsyncDisposable
     {
         private readonly TaskCompletionSource<object?> _watchingTcs = new TaskCompletionSource<object?>();
         private readonly PlcDataMapper _mapper;
         private readonly LruCache _lruCache = new LruCache();
         private readonly Dictionary<string, PlcWatchReference> _variables = new Dictionary<string, PlcWatchReference>();
-        private readonly SemaphoreSlim _lock;
+        private readonly SemaphoreSlim _semaphore;
         private readonly ChangeDetectionStrategy _changeDetectionStrategy;
         private readonly int _defaultInterval;
         private readonly AsyncAutoResetEvent<IEnumerable<DataPack>>? _changeEvent;
+        private readonly AsyncAutoResetEvent<bool> _modifiedEvent = new AsyncAutoResetEvent<bool>();
+        private bool _disposed;
 
         private CancellationTokenSource? _cts;
         private List<Execution>? _executions;
@@ -65,7 +67,7 @@ namespace Papper.Extensions.Notification
             _mapper = mapper ?? ExceptionThrowHelper.ThrowArgumentNullException<PlcDataMapper>(nameof(mapper));
             _changeDetectionStrategy = changeDetectionStrategy;
             _defaultInterval = defaultInterval;
-            _lock = new SemaphoreSlim(1);
+            _semaphore = new SemaphoreSlim(1);
             UpdateWatchCycle(vars);
 
             if (changeDetectionStrategy == ChangeDetectionStrategy.Event)
@@ -74,7 +76,11 @@ namespace Papper.Extensions.Notification
             }
             if (vars != null)
             {
-                AddItems(vars);
+                // Add Items synchrone
+                if(TryBuilAddableVariableList(vars, out var variables, true))
+                {
+                    AddValidatedVariables(variables!);
+                }
             }
         }
 
@@ -83,45 +89,67 @@ namespace Papper.Extensions.Notification
         /// </summary>
         public void Dispose()
         {
-            using (new SemaphoreGuard(_lock))
-            { 
+            if (_disposed) return;
+            using (new SemaphoreGuard(_semaphore))
+            {
                 _watchingTcs.TrySetResult(null);
                 _mapper.RemoveSubscription(this);
                 _variables.Clear();
                 _lruCache.Dispose();
             }
-            _lock.Dispose();
+            _semaphore.Dispose();
+            _disposed = true;
+        }
+
+        /// <summary>
+        /// Disposing all allocations and unregister from mapper
+        /// </summary>
+        public async ValueTask DisposeAsync()
+        {
+            if (_disposed) return;
+            if (_semaphore != null)
+            {
+                using (await SemaphoreGuard.Async(_semaphore).ConfigureAwait(false))
+                {
+                    await InternalStopSubscription().ConfigureAwait(false);
+                }
+                _semaphore.Dispose();
+                _disposed = true;
+            }
         }
 
         /// <summary>
         /// This method cancels the current call of DetectChangesAsync. After calling the DetectChangesAsync again,
         /// this cancellation is automatically reseted and ready for a newly cancellation.
         /// </summary>
-        public void CancelCurrentDetection() => _cts?.Cancel();
+        public void CancelCurrentDetection()
+        {
+            _cts?.Cancel();
+        }
 
         /// <summary>
         /// Add items to the subscription. This items are active in the next watch cycle, so you will get a data change if the update was activated.
         /// </summary>
         /// <param name="vars">Vars to activate </param>
-        public void AddItems(params PlcWatchReference[] vars) => AddItems(vars as IEnumerable<PlcWatchReference>);
+        public Task<bool> AddItemsAsync(params PlcWatchReference[] vars) => AddItemsAsync(vars as IEnumerable<PlcWatchReference>);
 
         /// <summary>
         /// Add items to the subscription. This items are active in the next watch cycle, so you will get a data change if the update was activated.
         /// </summary>
         /// <param name="vars">Vars to activate </param>
-        public bool TryAddItems(params PlcWatchReference[] vars) => TryAddItems(vars as IEnumerable<PlcWatchReference>);
+        public Task<bool> TryAddItemsAsync(params PlcWatchReference[] vars) => TryAddItemsAsync(vars as IEnumerable<PlcWatchReference>);
 
         /// <summary>
         /// Add items to the subscription. This items are active in the next watch cycle, so you will get a data change if the update was activated.
         /// </summary>
         /// <param name="vars">Vars to activate </param>
-        public void AddItems(IEnumerable<PlcWatchReference> vars) => InternalAddItems(vars, true);
+        public Task<bool> AddItemsAsync(IEnumerable<PlcWatchReference> vars) => InternalAddItemsAsync(vars, true);
 
         /// <summary>
         /// Add items to the subscription. This items are active in the next watch cycle, so you will get a data change if the update was activated.
         /// </summary>
         /// <param name="vars">Vars to activate </param>
-        public bool TryAddItems(IEnumerable<PlcWatchReference> vars) => InternalAddItems(vars, false);
+        public Task<bool> TryAddItemsAsync(IEnumerable<PlcWatchReference> vars) => InternalAddItemsAsync(vars, false);
 
 
         /// <summary>
@@ -129,17 +157,17 @@ namespace Papper.Extensions.Notification
         /// The internal 
         /// </summary>
         /// <param name="vars"></param>
-        public bool RemoveItems(params PlcWatchReference[] vars) => RemoveItems(vars as IEnumerable<PlcWatchReference>);
+        public Task<bool> RemoveItemsAsync(params PlcWatchReference[] vars) => RemoveItemsAsync(vars as IEnumerable<PlcWatchReference>);
 
         /// <summary>
         /// Remove items from the watch list. This items will be removed before the next watch cycle.
         /// The internal 
         /// </summary>
         /// <param name="vars"></param>
-        public bool RemoveItems(IEnumerable<PlcWatchReference> vars)
+        public async Task<bool> RemoveItemsAsync(IEnumerable<PlcWatchReference> vars)
         {
             var filteredVars = vars.Where(item => !string.IsNullOrEmpty(item.Address)).ToList();
-            using (new SemaphoreGuard(_lock))
+            using (await SemaphoreGuard.Async(_semaphore).ConfigureAwait(false))
             {
                 var result = filteredVars.Any(item => _variables.Remove(item.Address)) | _modified;
                 if (result)
@@ -147,6 +175,7 @@ namespace Papper.Extensions.Notification
                     UpdateWatchCycle(_variables.Values);
                 }
                 _modified = result;
+                _modifiedEvent?.Set(true);
                 return result;
             }
         }
@@ -178,17 +207,17 @@ namespace Papper.Extensions.Notification
                 return null;
             }
 
-            var variables = vars.Select(x => x.Address).ToList();
             List<Execution>? executions = null;
-            using (await SemaphoreGuard.Async(_lock))
+            using (await SemaphoreGuard.Async(_semaphore).ConfigureAwait(false))
             {
-                executions = _executions!.ToList();
-                if (executions!.IsNullOrEmpty())
-                {
-                    return null;
-                }
+                executions = _executions?.ToList();
             }
 
+            if (executions!.IsNullOrEmpty())
+            {
+                return null;
+            }
+            var variables = vars.Select(x => x.Address).ToList();
             return executions.GroupBy(exec => exec.ExecutionResult) // Group by execution result
                                                   .SelectMany(group => group.SelectMany(g => g.Bindings)
                                                                             .Where(b => variables.Contains(b.Key))
@@ -205,6 +234,7 @@ namespace Papper.Extensions.Notification
         /// <returns></returns>
         public async Task<ChangeResult> DetectChangesAsync()
         {
+
             var cts = new CancellationTokenSource();
             if (Interlocked.CompareExchange(ref _cts, cts, null) != null)
             {
@@ -226,7 +256,7 @@ namespace Papper.Extensions.Notification
                     if (_modified)
                     {
                         // test if sema is free, if not test it the next cycle
-                        if (_lock.Wait(0))
+                        if (_semaphore.Wait(0))
                         {
                             try
                             {
@@ -245,7 +275,7 @@ namespace Papper.Extensions.Notification
                             }
                             finally
                             {
-                                _lock.Release();
+                                _semaphore.Release();
                             }
                         }
                     }
@@ -273,7 +303,7 @@ namespace Papper.Extensions.Notification
                         }
                         else
                         {
-                            var packs = await _changeEvent.WaitAsync().ConfigureAwait(false);
+                            var packs = await _changeEvent.WaitAsync(_cts.Token).ConfigureAwait(false);
                             if (packs != null && packs.Any())
                             {
                                 readRes = PlcDataMapper.CreatePlcReadResults(_executions, packs);
@@ -294,9 +324,18 @@ namespace Papper.Extensions.Notification
 
                     try
                     {
-                        await Task.Delay(interval, _cts.Token).ConfigureAwait(false);
+                        using (var timeoutCts = new CancellationTokenSource(TimeSpan.FromMilliseconds(interval)))
+                        {
+                            using (var linkedCts = CancellationTokenSource.CreateLinkedTokenSource(cts.Token, timeoutCts.Token))
+                            {
+                                //await Task.Delay(interval, linkedCts.Token).ConfigureAwait(false);
+                                await _modifiedEvent.WaitAsync(linkedCts.Token).ConfigureAwait(false);
+                            }
+                        }
                     }
-                    catch (TaskCanceledException) { }
+                    catch (TaskCanceledException) 
+                    { 
+                    }
 
                     if (_cts.IsCancellationRequested)
                     {
@@ -307,8 +346,8 @@ namespace Papper.Extensions.Notification
             }
             finally
             {
-                _cts.Dispose();
                 _cts = null;
+                cts?.Dispose();
             }
         }
 
@@ -323,6 +362,16 @@ namespace Papper.Extensions.Notification
                                           !cycles.Any() &&
                                           !string.IsNullOrWhiteSpace(variables.FirstOrDefault(x => cycles.TryGetValue(x, out var interval) &&
                                                                                                    lastChange.AddMilliseconds(interval < cycleInterval * 2 ? cycleInterval : interval) < DateTime.Now));
+        private async Task InternalStopSubscription()
+        {
+            _watchingTcs.TrySetResult(null);
+            await _watchingTcs.Task.ConfigureAwait(false);
+            _mapper.RemoveSubscription(this);
+            _variables.Clear();
+            _lruCache.Dispose();
+        }
+
+
 
         internal void OnDataChanged(IEnumerable<DataPack> changed)
         {
@@ -336,15 +385,27 @@ namespace Papper.Extensions.Notification
             }
         }
 
-        private bool InternalAddItems(IEnumerable<PlcWatchReference>? vars, bool throwExceptions)
+        private async Task<bool> InternalAddItemsAsync(IEnumerable<PlcWatchReference>? vars, bool throwExceptions)
+        {
+            if (TryBuilAddableVariableList(vars, out var variables, throwExceptions))
+            {
+                using (await SemaphoreGuard.Async(_semaphore).ConfigureAwait(false))
+                {
+                    return AddValidatedVariables(variables!);
+                }
+            }
+            return false;
+        }
+
+        private bool TryBuilAddableVariableList(IEnumerable<PlcWatchReference>? vars, out Dictionary<string, PlcWatchReference>? variables,  bool throwExceptions)
         {
             if (vars == null || !vars.Any())
             {
+                variables = null;
                 return false;
             }
 
-
-            var variables = new Dictionary<string, PlcWatchReference>();
+            variables = new Dictionary<string, PlcWatchReference>();
             foreach (var variable in vars)
             {
                 if (!_mapper.IsValidReference(variable))
@@ -357,31 +418,28 @@ namespace Papper.Extensions.Notification
                 }
                 variables.Add(variable.Address, variable);
             }
+            return variables.Any();
+        }
 
-            if (variables.Any())
+        private bool AddValidatedVariables(Dictionary<string, PlcWatchReference> variables)
+        {
+            // After we validate all items, we add all
+            // if one is not valid non of them will be added, this is because of consistence
+            foreach (var variable in variables.Values)
             {
-                using (new SemaphoreGuard(_lock))
+                if (_variables.ContainsKey(variable.Address))
                 {
-                    // After we validate all items, we add all
-                    // if one is not valid non of them will be added, this is because of consistence
-                    foreach (var variable in variables.Values)
-                    {
-                        if (_variables.ContainsKey(variable.Address))
-                        {
-                            _variables[variable.Address] = variable;
-                        }
-                        else
-                        {
-                            _variables.Add(variable.Address, variable);
-                        }
-                    }
-
-
-                    UpdateWatchCycle(_variables.Values);
-                    return _modified = true;
+                    _variables[variable.Address] = variable;
+                }
+                else
+                {
+                    _variables.Add(variable.Address, variable);
                 }
             }
-            return false;
+            UpdateWatchCycle(_variables.Values);
+            var result = _modified = true;
+            _modifiedEvent?.Set(true);
+            return result;
         }
